+++ lib/x/client.rb
@@ -1,3 +1,4 @@
+require "forwardable"
 require "json"
 require "net/http"
 require "oauth"
@@ -16,6 +17,12 @@ module X
 
   # HTTP client that handles authentication and requests
   class Client
+    extend Forwardable
+    def_delegators :@http_request, :bearer_token, :bearer_token=, :api_key,
+                   :api_key=, :api_key_secret, :api_key_secret=, :access_token, :access_token=,
+                   :access_token_secret, :access_token_secret=,
+                   :base_url, :base_url=, :user_agent, :user_agent=
+
     DEFAULT_BASE_URL = "https://api.twitter.com/2/".freeze
     DEFAULT_USER_AGENT = "X-Client/#{X::Version} Ruby/#{RUBY_VERSION}".freeze
 
@@ -55,6 +62,19 @@ module X
 
     # HTTP client requester
     class HttpRequest
+      extend Forwardable
+      attr_accessor :base_url, :user_agent
+      attr_reader :bearer_token
+
+      def_delegator :@access_token, :secret, :access_token_secret
+      def_delegator :@access_token, :secret=, :access_token_secret=
+      def_delegator :@access_token, :token, :access_token
+      def_delegator :@access_token, :token=, :access_token=
+      def_delegator :@consumer, :key, :api_key
+      def_delegator :@consumer, :key=, :api_key=
+      def_delegator :@consumer, :secret, :api_key_secret
+      def_delegator :@consumer, :secret=, :api_key_secret=
+
       HTTP_METHODS = {
         get: Net::HTTP::Get,
         post: Net::HTTP::Post,
@@ -89,6 +109,11 @@ module X
 
       def delete(endpoint)
         send_request(:delete, endpoint)
+      end
+
+      def bearer_token=(bearer_token)
+        @use_bearer_token = !bearer_token.nil?
+        @bearer_token = bearer_token
       end
 
       private